If disableSubmit is true, trigger validation on setup.
Browsers will autocomplete POST form data. If disableSubmit is true, and
a browser autofills (correct) data, the validateField function will not
fire, and the user will have to manually blur each field in order to
submit. This is undesirable from a UX perspective.

If disableSubmit is true, validateField is bound to the oninput event,
which will trigger during autocompletion. Since the browser can
autocomplete fields before JS loads, the event is then triggered.